+++ packages/miniplex-core/test/Bucket.test.ts
@@ -185,6 +185,80 @@ describe("derive", () => {
   })
 })
 
+describe("clear", () => {
+  it("removes all entities from the bucket", () => {
+    const bucket = new Bucket()
+    bucket.add({ count: 1 })
+    bucket.add({ count: 2 })
+    expect(bucket.entities).toEqual([{ count: 1 }, { count: 2 }])
+
+    bucket.clear()
+    expect(bucket.entities).toEqual([])
+  })
+
+  it("emits an event for each entity that is removed", () => {
+    const bucket = new Bucket<{ count: number }>()
+    const listener = jest.fn()
+    bucket.onEntityRemoved.addListener(listener)
+
+    bucket.add({ count: 1 })
+    bucket.add({ count: 2 })
+    bucket.clear()
+    expect(listener).toHaveBeenCalledTimes(2)
+  })
+
+  it("emits the onCleared event", () => {
+    const bucket = new Bucket()
+    const listener = jest.fn()
+    bucket.onCleared.addListener(listener)
+
+    bucket.clear()
+    expect(listener).toHaveBeenCalledTimes(1)
+  })
+})
+
+describe("dispose", () => {
+  it("removes all entities from the bucket", () => {
+    const bucket = new Bucket()
+    bucket.add({ count: 1 })
+    bucket.add({ count: 2 })
+    expect(bucket.entities).toEqual([{ count: 1 }, { count: 2 }])
+
+    bucket.dispose()
+    expect(bucket.entities).toEqual([])
+  })
+
+  it("also disposes any derived buckets", () => {
+    const bucket = new Bucket()
+    const derivedBucket = bucket.derive()
+    bucket.add({ count: 1 })
+    expect(derivedBucket.entities).toEqual([{ count: 1 }])
+
+    bucket.dispose()
+    expect(derivedBucket.entities).toEqual([])
+  })
+
+  it("also disposes buckets derived from derived buckets", () => {
+    const bucket = new Bucket()
+    const derivedBucket = bucket.derive()
+    const derivedBucket2 = derivedBucket.derive()
+    bucket.add({ count: 1 })
+    expect(derivedBucket2.entities).toEqual([{ count: 1 }])
+
+    bucket.dispose()
+    expect(derivedBucket2.entities).toEqual([])
+  })
+
+  it("when a derived bucket is disposed, remove its listeners from us", () => {
+    const bucket = new Bucket()
+    const derivedBucket = bucket.derive()
+    expect(bucket.onEntityAdded.listeners.size).toEqual(1)
+
+    derivedBucket.dispose()
+    expect(bucket.onEntityAdded.listeners.size).toEqual(0)
+  })
+})
+
 describe("size", () => {
   it("returns the size of the world", () => {
     const bucket = new Bucket()