+++ packages/miniplex-react/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 2.0.0-next.6
+
+### Patch Changes
+
+- 3fcc79a: When `<Property>` encounters an entity that has the property already set, it will not add or remove the proprty on mount or unmount, but it will still update the property if the value changes. When unmounting, it will restore the property to the value it had before the component was mounted.
+
 ## 2.0.0-next.5
 
 ### Patch Changes